[NFC] Refactor STPSolver and STPSolverImpl out of Solver.cpp into their own file ``STPSolver.cpp``. Whilst I'm here also clang-format the modified code.
